Updated examples slightly. Added physical properties quantities (for mass).
Considering merging PhysicalProperties into Physics since there's some interrelations going on there.

Notes on things to look at:

Unitals tend to be example specific, so they are not easily pulled out (one exception being mass, which has not yet been extracted fully), however, after looking at it for a bit the idea of "parameterized unitals" would seem to be a good approach. Ex: mass of water would be a unital derived from mass, with water parameters fed in. Mass of PCM would then be handled in the same way (using PCM instead of Water).

Not sure if we should have "Unitals" namespaces in the same way as concepts and quantities, because there would most likely be very few of them, (at least, prior to parameterization).