Convert all of operators to new method using DomainDescr.
In doing so, needed to remove some of the weirder cases of "bounds" that
made no sense. Removed some hacks wrt printing of integrals -- but
mostly moved them down further into the printing code, rather than have
them further up.

Change the AST for HTML and TeX to have Symbol for the variable.

And adjust all the examples to match.  Prints exactly as before!