Another two-for-one commit.
These two changes needed to be done together:
- remove Chunk/Wrapper/UWrapper
- add Chunk/UnitaryConcept

But the implementation was awkward, and I knew better could be done.
So I split up Concept into Definition and ConceptDomain. This needed
quite a few fixes in a lot of places, but made quite a bit of code
much more elegant. I could have (should have) done that change on its
own first.  Oh well.

As far as I know, this is the last of the Wrappers.