Rather large re-organization of things to do with 'Symbols'. There was lots of stuff in SymbolForm that made no sense to me (some of it because of historical reasons), and then some really similar stuff embedded in Quantity.
In Quantity, there were two ways to get at a Symbol: symbol (a total function which
used Stage) and getStagedS (a partial function, implented as a Map). They were
supposed to be the same, but this is dangerous.  From other parts in the code, it
was clear that this function needed to be partial [somehow]. The 'simplest' way to do
this would be to add a new Symbol, Empty, which represents the absence of Symbol.
Empty should not be used (even visible) outside of Drasil. It is only used internally
at the top-level of a Symbol to signal 'absence'.

So, overall:
- symbol (was getSymb) is no longer part of Quantity, but of HasSymbol
- Quantity demands HasSymbol
- getStagedS from Quantity is gone
- eqSymb and codeSymb have both moved to Chunk.SymbolForm
- this made some instances 'orphaned', which were fixed by putting them in the right place
- SymbolForm now exports just
  HasSymbol(..), Stage(..), eqSymb, codeSymb, hasStageSymbol
- in particular, StagedSymbolChunk is gone, and all its many constructors
- StagedSymbolChunk uses are replaced with (Stage -> Symbol) functions
- the addition of Empty makes Symbol into a Monoid -- instance implemented (not yet used)
- TeX printing needs to know what the 'current context' is needed for Empty.
  "Easy": it doesn't care. Problem is, that was not (yet) part of the MathContext lattice.
  It now is -- it is called 'Curr' (for current).
- this enabled some slight simplifications of functions in DocumentLanguage and TableOfSymbols.