Created a test for sorting interleaved inputs.
The general heuristic is to greedily add lines, with constraint-checking
having highest priority, then derived values, and parsing only when the
first two cannot be done.  I also chose to sort constraint-checks and
derived values based on number of dependencies, so that `a < 10` will
always take place before `a < b`, if `a` and `b` are both defined.