You'd better believe I found another edge case
in the `listSlice` from `Macros.hs`.  If `step` is a variable, and the order of `beg` and `end` is opposite of
the sign of `step`'s value, then the slice will contain `1` element
rather than `0` like it should.  This new test case checks for that
case.

The fix will be pretty easy, coming soon.