Changed the sorting approach to be more concise.
I created a new datatype to store the kind of statement, and used it and
the number of dependencies to naively sort the statements by the
heuristics before running topological sort.  Once that is done, we can
just repeatedly choose the first element that has all of its
dependencies met.  This should actually produce slightly better results
(though they're the same for these tests), and is hopefully a bit more
concise and readable.