Split `ClassInterface.hs` into separate files.
I called them `InterfaceCommon.hs` and `InterfaceGOOL.hs`.  I chose to
postfix the `Shared` and `GOOL` parts so that the files are together
when sorted alphabetically.

I tried not to make any changes to the interface itself; however I did
notice one method I had left in `ValueExpression`, and I moved it to
`OOValueExpression` where it belongs.