Removed method calls from `listAccess` and `listSet`
Both `listAccess` and `listSet` are implemented as a method call in the
implementations used by all targets.  I thought this  was weird, as
they are only actually method calls for Java and C++.  I changed the
implementation to create a value without worrying what kind of call it
is.

Another option would be to split these functions up into functions that
use methods and functions that don't.  I opted to keep them together, as
`listAccess` in particular is large enough that I'd rather not duplicate
it, when this way does essentially the same thing.