Separated more typeclasses into OO and Procedural
I made parallel versions of `ModuleSym`, `FileSym`, and `ProgramSym`
(along with their respective associated types).  Because `ModuleSym` was
the only thing relying on `ClassSym`, `ClassSym` is now fully-OO; so I
moved it into the OO file.

Note that while Procedural versions of `ModuleSym`, `FileSym`, and
`ProgramSym` exist, nothing uses them, and `GOOL.Drasil` doesn't export
them.  The latter is because they have the same name, and thus need to
be imported with a qualified name if both versions of the typeclasses
are used in the same file.

Builds on #3849