Separated a few more GOOL typeclasses
- Created `OOMethodSym` to hold the OO-only parts of `MethodSym`.
    - I don't like the name.  Ideally we'd split `MethodSym` into `FunctionSym` and `MethodSym`.  Currently `FunctionSym` is taken, but I don't like the name there so maybe at some point we can do a 3-way rename.
- Moved `StateVarSym` and `PermanenceSym` into `InterfaceGOOL`, as all of their dependencies are now in there.

My diagram tells me that the only OO typeclass left in `InterfaceCommon`
is `ScopeSym`, which is still required by `MethodSym` for some reason.
I'll look into how those functions are implemented, and see if we can
get rid of it.  Otherwise we'll need to split `MethodSym` completely
into `OOMethodSym` and `ProcMethodSym`.  We'll see what works.

Oh, and I also got rid of a bunch of whitespace at the end of the line
in the C++ renderer.  Probably should have done that in a different PR 😅.

Builds on #3850