Split `GOOL.Drasil` into two separate files
They are now `Drasil.GOOL` and `Drasil.GProc`.  This will allow us to
differentiate between OO and Procedural languages in `drasil-code`.

We had talked about splitting these into separate packages of Drasil,
but after talking with @balacij, I decided not to do that just yet.
There's not much that we lose by doing it later, and I'm not sure if we
know well enough at this point how separate they should be.  Right now, for example, the
only files that are purely Procedural is `GProc.hs` and
`InterfaceProc.hs`, and only a few more than that which are purely OO.
There will likely be more, but I doubt it'll be a ton.  Since there is
so much code that is shared, we might just want to keep them in the same
package.