Fixed `listSlice` in Julia
That was quite a bit more complicated than I expected.  Julia is
1-indexed, but `end` is inclusive in its list slices, which means we
don't add 1 if `step` is positive, but we need to add 2 if `step` is
negative.  Also, if `beg` or `end` aren't given but `step` is a literal,
then we can use the `begin` or `end` keyword instead of `1` or
`size(list)`.

I also fixed the other renderers to make sure they generate fresh names
for `begIdx` and `endIdx`, as I realized that if two list slices
required one of those, there would be a name conflict.

I also added the `smartAdd` and `smartSub` functions to
`LanguagePolymorphic.hs`, which do basic constant-folding.  I
thought about modifying `(#+)` and `(#-)` directly, but I figured there
might be cases where we don't want constant-folding.