Made `staticVar` uppercase for Python constants
I added an extra boolean parameter to `staticVar` (disregarded by all
but the Python renderer) to say if the static variable is a variable or
a constant.  Then I used that to uppercase static constants in the
Python renderer.

The one thing I don't like about this is that it requires static
variables to be slightly dishonest about their contents: their name is
as given, while their contents are uppercase.