Removed `scopeDoc` from `ScopeData`
Initially I thought it might be handy in the future to have the added
flexibility of using a different keyword for `global`/`local`
declarations, but now I'm thinking that might be a little bit too
premature.  Only one of our renderers uses `ScopeSym`, and when we add
another one we don't know what will be handy to have in `ScopeData`.  We
can discuss this more in the PR.