Added getting one variable's scope for another one
There are a few macros that involve creating a temporary variable.
Previously I just set them to be local, but now they use the scope of
the input variable to set their own scope.

This has a possibility of being `global` more often than it should, but
it should be more accurate than it was.