Updated range in Julia
Julia's ranges include the end value, and other languages don't. I
updated the Julia renderer to subtract `1` from the end value to match
the other languages.

Note that while this fixes it for positive `step` values, it does not
for negative `step` values. More work needs to be done here.