Remove unused ConfigParser in drasil-code.
Closes #2469

The `ConfigParser.hs` file is both stale, largely incomplete, and tied
to existing `Choices`, which we are going to re-work. If/when
we de-embed data from Drasil, this file would need to be re-designed
anyways, and right now, keeping it in the `drasil-code` folder means we
need to dance around it in `package.yaml`. So, I think we can remove it
for now in favour of making `drasil-code`'s `package.yaml` file a bit
nicer.